Add support to different MTU values for internal and external IPv4 routes
Previously, there is no way to set up different MTU values for internal
routes (same cluster) and external routes, including to other nodes in
the clustermesh. So it was impossible to configure the internal network
to use Jumbo Frames and keep the clustermesh communication working on a
multi-region setup that relies, for example, on AWS VPC peering
connection, limited to 1500 bytes.

Now it is possible to use the --external-mtu option to configure a
custom MTU value for external routing.